backend/requirements.txt
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asyncio==3.4.3
attrs==25.1.0
boto3>=1.28.0
dnspython==2.7.0
fastapi==0.115.11
frozenlist==1.5.0
httpx>=0.26.0
idna==3.10
langchain>=0.1.0
langchain-aws>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.10
langchain-mongodb>=0.0.1
multidict==6.1.0
propcache==0.3.0
pydantic==2.10.6
pydantic_core==2.27.2
pymongo==4.11.2
sniffio==1.3.1
sseclient-py==1.7.2
starlette==0.46.0
typing_extensions==4.12.2
yarl==1.18.3
